Handle target versions while deleting content blocks

+++ app/mutators/delete_content_block_mutator.rb
@@ -5,7 +5,17 @@ class DeleteContentBlockMutator < ApplicationMutator
   validate :not_the_only_content_block
 
   def delete_content_block
-    content_block.destroy!
+    ContentBlock.transaction do
+      if latest_content_version.updated_at.to_date == Date.today
+        content_block.destroy!
+        latest_content_version.content_blocks -= [id.to_i]
+        latest_content_version.save!
+      else
+        new_version = latest_content_version.dup
+        new_version.content_blocks -= [id.to_i]
+        new_version.save!
+      end
+    end
   end
 
   def not_the_only_content_block
@@ -22,5 +32,13 @@ class DeleteContentBlockMutator < ApplicationMutator
 
   def authorized?
     current_school_admin.present? || current_user.course_authors.where(course: content_block.target.level.course).exists?
+  end
+
+  def target
+    @target ||= content_block.target
+  end
+
+  def latest_content_version
+    @latest_content_version ||= target.target_content_versions.order('updated_at desc').first
   end
 end
